+++ src/core/transactions/transactions/result/state/TransactionState.h
@@ -24,7 +24,7 @@ public:
     static TransactionState::SharedConst flushAndContinue();
     static TransactionState::SharedConst awakeAsFastAsPossible();
     static TransactionState::SharedConst awakeAfterMilliseconds(
-        uint16_t milliseconds);
+        uint32_t milliseconds);
 
     static TransactionState::SharedConst waitForMessageTypes(
         vector<Message::MessageTypeID> &&requiredMessageType,
